\name{genBinomTest-package}
%% \alias{genBinomTest-package}
%% \alias{genBinomTest}
\docType{package}
\title{
Generalized binomial test
}
\description{
\code{genBinomTest} provides an implementation of a generalized binomial
test for independent Bernoulli random variables.
}
\details{
\tabular{ll}{
Package: \tab genBinomTest\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-10-14\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Karl Schlag and Oliver Reiter
}
\references{
  Schlag, K. (2013), "Doing Hoeffding's Homework - Tests and Confidence
  Intervals for the mean of finitely many independently distributed
  Bernoulli random variables."
  
  \url{http://homepage.univie.ac.at/karl.schlag/research/statistics/}
 
}

